+++ addOns/client/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Wait always after component navigation while crawling.
+- Maintenance changes.
 
 ## [0.29.0] - 2026-06-30
 ### Changed