+++ addOns/client/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Wait always after component navigation while crawling.
 
+### Fixed
+- Normalise behaviour of Delete context menu item.
+
 ## [0.29.0] - 2026-06-30
 ### Changed
 - Use new functionality from the browser extension for crawling.